Add files via upload
This project is a responsive resume webpage created using HTML and CSS. It showcases my professional background, skills, education, and project experience in a clean and modern layout. The design is fully customizable and adaptable to different devices, ensuring a seamless user experience. The project demonstrates my proficiency in front-end web development and attention to detail.